test: isolate common steps
