+++ tests/e2e-tests/NHS.CohortManager.EndToEndTests/Steps/Common_AutomatedRegressionSuiteSteps.cs
@@ -6,31 +6,35 @@ using Microsoft.Extensions.DependencyInjection;
 using FluentAssertions;
 using NHS.CohortManager.EndToEndTests.Config;
 using NHS.CohortManager.EndToEndTests.Contexts;
+using NHS.CohortManager.EndToEndTests;
 using NHS.CohortManager.EndToEndTests.Models;
 using Microsoft.Extensions.Logging;
 using System;
 using System.Threading.Tasks;
 using System.Linq;
 using System.IO;
-
-
+using NHS.CohortManager.EndToEndTests.Helpers;
+using NUnit.Framework.Internal;
 
 [Binding]
 public class Common_AutomatedRegressionSuiteSteps
 {
     private readonly EndToEndFileUploadService _fileUploadService;
 
+    private readonly BlobStorageHelper _blobStorageHelper;
+
     private readonly AppSettings _appSettings;
     private EndToEndTestsContext _endtoendTestsContext;
 
 
-    public Common_AutomatedRegressionSuiteSteps(IServiceProvider services, AppSettings appSettings, EndToEndTestsContext endtoendTestsContext, ILogger<Epic1_AutomatedRegressionSuiteSteps> logger)
+    public Common_AutomatedRegressionSuiteSteps(IServiceProvider services, AppSettings appSettings, EndToEndTestsContext endtoendTestsContext, BlobStorageHelper blobStorageHelper)
     {
         _appSettings = appSettings;
         _endtoendTestsContext = endtoendTestsContext;
         _fileUploadService = services.GetRequiredService<EndToEndFileUploadService>();
-
+        _blobStorageHelper = blobStorageHelper;
     }
+
 
     [Given(@"the database is cleaned of all records for NHS Numbers: (.*)")]
     public async Task GivenDatabaseIsCleaned(string nhsNumbersString)
@@ -46,6 +50,7 @@ public class Common_AutomatedRegressionSuiteSteps
     {
         _fileUploadService.Should().NotBeNull("EndToEndFileUploadService is not initialized.");
     }
+
 
     [Given(@"file (.*) exists in the configured location for ""(.*)"" with NHS numbers : (.*)")]
     public void GivenFileExistsAtConfiguredPath(string fileName, string? recordType, string nhsNumbersData)
@@ -67,10 +72,23 @@ public class Common_AutomatedRegressionSuiteSteps
         await _fileUploadService.UploadFileAsync(filePath);
     }
 
+
     [Given(@"the NHS numbers in the database should match the file data")]
     [Then(@"the NHS numbers in the database should match the file data")]
     public async Task ThenVerifyNhsNumbersInDatabase()
     {
         await _fileUploadService.VerifyNhsNumbersAsync("BS_COHORT_DISTRIBUTION", _endtoendTestsContext.NhsNumbers!);
     }
+
+
+    [Then(@"verify the NhsNumbers in Participant_Management table should match (.*)")]
+    public async Task ThenVerifyTheInParticipantManagementShouldMatchAmended(string expectedRecordType)
+    {
+
+        await _fileUploadService.VerifyNhsNumbersAsync(
+            "PARTICIPANT_MANAGEMENT",
+            _endtoendTestsContext.NhsNumbers!,
+            expectedRecordType.ToUpper());
+    }
+
 }
